Preprocess_Data Code Added
I've managed to lop off one of the duplicate loops via a code detour. Given that, it seems to work and thus I deemed it fit to put in the proposal now. In addition, I forgot to dis-echo and suppress warnings on Vishal's data importing code when I put an R block around it.