Update ProjectProposal.Rmd
Just a heads up for anyone putting graphs in (you might already know this but I wasn't aware of this so I figured I'd tell everyone just in case).  If you want to insert a graph, put in the code that generates it instead of ggsaving it to the img folder.