Retooled Preliminary Analysis
I just noticed that in our preliminary analysis, one of the graphs uses efficiency before the efficiency metric is derived via r value. While we do briefly mention how efficiency is calculated in the analysis plan, I do believe we should derive efficiency in our prelim analysis before we use it in any graphs.

Very sorry for  the inconvenience, I did not notice this until I was reviewing the proposal this morning.