Added Data Cleaning and Importing
I have added the following R code to our draft report:
- Data Import and Clean for main data set
- Air traffic Data Import and Format
- Weather Data Import and Format
I also added a bit of the analysis from our proposal, namely the derivation and graphing of the efficiency metric. I hope to, by the end of today,  determine the ideal graph sizing such that two graphs can fit on the same line and both still be readable and properly formatted.